Move tests into src directory into their respective __tests__ folders
